+++ ontimize/config/o-modules.ts
@@ -65,7 +65,7 @@ export const ONTIMIZE_MODULES: any = [
   // Ng2-translate
   TranslateModule.forRoot({
     provide: TranslateLoader,
-    useFactory: (http: Http) => new TranslateStaticLoader(http, '/assets/i18n', '.json'),
+    useFactory: (http: Http) => new TranslateStaticLoader(http, './assets/i18n', '.json'),
     deps: [Http]
   }),